Extracted page cell tests to a different file
