+++ GiniVision/Classes/Core/Multipage Review/MultipageReviewViewController.swift
@@ -494,6 +494,7 @@ extension MultipageReviewViewController: UICollectionViewDataSource {
             let cell = collectionView
                 .dequeueReusableCell(withReuseIdentifier: MultipageReviewPagesCollectionCell.identifier,
                                      for: indexPath) as? MultipageReviewPagesCollectionCell
+
             cell?.setUp(with: documentRequests[indexPath.row], at: indexPath.row, giniConfiguration: giniConfiguration)
             return cell!
         }
